entities base_layout: extract the actions menu to a sub-component
